worker/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
faster-whisper==1.0.3
pyannote.audio==3.3.2
torch==2.2.0
torchaudio==2.2.0
requests==2.31.0
python-multipart==0.0.9
